+++ dask_geomodeling/raster/temporal.py
@@ -8,6 +8,7 @@ from pandas.tseries.frequencies import to_offset
 
 import numpy as np
 import pandas as pd
+import warnings
 
 from dask_geomodeling.utils import (
     get_dtype_max,
@@ -245,7 +246,7 @@ def _ts_to_dt(timestamp, timezone):
         timestamp = timestamp.tz_localize(timezone)
     except TypeError:
         pass
-    return timestamp.tz_convert("UTC").tz_localize(None).to_pydatetime()
+    return timestamp.tz_convert("UTC").tz_localize(None).to_pydatetime(warn=False)
 
 
 def _get_bin_label(dt, frequency, closed, label, timezone):
@@ -292,7 +293,7 @@ def _get_closest_label(dt, frequency, closed, label, timezone, side="both"):
         differences = differences[differences >= pd.Timedelta(0)]
     elif side == "left":
         differences = differences[differences <= pd.Timedelta(0)]
-    result = differences.abs().argmin()
+    result = differences.abs().idxmin()
     return _ts_to_dt(result, timezone)
 
 
@@ -566,7 +567,9 @@ class TemporalAggregate(BaseSingle):
         times = time_data["time"]
 
         # convert times to a pandas series
-        series = pd.Series(index=times).tz_localize("UTC").tz_convert(timezone)
+        series = (
+            pd.Series(index=times, dtype=float).tz_localize("UTC").tz_convert(timezone)
+        )
 
         # localize the labels so we can use it as an index
         labels = labels.tz_localize("UTC").tz_convert(timezone)
@@ -620,7 +623,10 @@ class TemporalAggregate(BaseSingle):
             inds = indices[timestamp]
             if len(inds) == 0:
                 continue
-            aggregated = agg_func(values[inds], axis=0)
+            with warnings.catch_warnings():
+                # the agg_func could give use 'All-NaN slice encountered'
+                warnings.simplefilter("ignore", category=RuntimeWarning)
+                aggregated = agg_func(values[inds], axis=0)
             # keep track of NaN or inf values before casting to target dtype
             no_data_mask = ~np.isfinite(aggregated)
             # cast to target dtype
@@ -785,7 +791,9 @@ class Cumulative(BaseSingle):
         closed = process_kwargs["closed"]
         label = process_kwargs["label"]
         times = (
-            pd.Series(index=time_data["time"]).tz_localize("UTC").tz_convert(timezone)
+            pd.Series(index=time_data["time"], dtype=float)
+            .tz_localize("UTC")
+            .tz_convert(timezone)
         )
 
         if frequency is None: