+++ CHANGES.rst
@@ -2,10 +2,21 @@ Changelog of dask-geomodeling
 ===================================================
 
 
-2.2.1 (unreleased)
+2.2.2 (unreleased)
 ------------------
 
 - Nothing changed yet.
+
+
+2.2.1 (2020-02-04)
+------------------
+
+- Suppressed "invalid value encountered in greater than" warning in
+  ClassifyFromColumns.
+
+- Compatibility fixes for pandas 1.0.0.
+
+- Implemented raster.RasterizeWKT
 
 
 2.2.0 (2019-12-20)